requirements.txt
# ======================
# Core / Env
# ======================
python-dotenv>=1.2.1
requests>=2.31.5
setuptools>=65.0.0

# ======================
# MongoDB
# ======================
pymongo>=4.15.5
dnspython>=1.16.0
certifi>=2023.11.17

# ======================
# Scraping
# ======================
selenium>=4.15.0
webdriver-manager>=4.0.1
beautifulsoup4==4.14.3

# ======================
# NLP / Data
# ======================
pandas>=2.0.0
scikit-learn>=1.3.0
nltk>=3.9.2
numpy>=2.4.0
scipy>=1.16.3
joblib>=1.5.3


# ======================
# NLP / Embeddings & Clustering
# ======================
sentence-transformers>=2.3.1
umap-learn>=0.5.6
numba>=0.60.0
hdbscan>=0.8.38


# ======================
# Visualisation
# ======================
networkx>=3.2
plotly>=5.18.0
matplotlib>=3.10.8
seaborn>=0.13.2

# ======================
# Excel
# ======================
openpyxl>=3.1.5
et_xmlfile>=2.0.0

# ======================
# App
# ======================
streamlit>=1.30.0
mistralai>=1.0.0

# ======================
# DuckDB
# ======================
duckdb>=0.10.0
